Bumped WPF-UI to 3.0.0
The most significant changes were made to the `Extensions.cs` file in the `MVVM_GMI.Helpers` namespace, the `API.cs` file in the `MVVM_GMI.Services.Database` namespace, and the `AuthWindowViewModel.cs` file in the `MVVM_GMI.ViewModels.Windows` namespace.

In the `Extensions.cs` file, the encryption method was changed from `RSACryptoServiceProvider` to `RSA.Create()`, and the encryption padding was changed from `false` to `RSAEncryptionPadding.Pkcs1`. This change was made to improve the security of the encryption process.

The `API.cs` file saw extensive modifications, including adjustments to error handling, changes to the way HTTP responses are processed, and updates to the way session tokens and access tokens are managed. These changes were made to improve the efficiency and reliability of the API.

In the `AuthWindowViewModel.cs` file, error handling during user registration and login was improved. The changes include more detailed error messages and adjustments to the conditions for showing certain windows. This was done to enhance the user experience and provide more informative feedback.

Other changes include:

- The `OnlineRequest.cs` file was updated to include additional `MessageBox.Show()` calls for debugging purposes and the formatting of the output messages was adjusted for better readability.
- The `MVVM_GMI.csproj` file was updated to change the versions of the `CmlLib.Core`, `MdXaml`, and `WPF-UI` packages.
- The `ApplicationHostService.cs` file was modified to change the condition for showing the navigation window. It now shows the window if the API response is successful, regardless of the user's membership status.
- The `SessionService.cs` file was updated to include a `MessageBox.Show()` call when the session token is updated.
- The `DashboardViewModel.cs` file was updated to include the `Wpf.Ui.Extensions` namespace.